Fixed a layout issue with IconRect.
Previously, it align the width by the height, and vice versa, which would produce odd results. In addition, it wouldn't shift the position of the returned Rect to the position the region was being calculated for, so it would default to the top middle of the given region, or the top left depending on the parent position.